# api/openapi.yaml
openapi: "3.0.3"
info:
  title: Sockerless Docker-Compatible API
  description: |
    Docker Engine API v1.44 subset implemented by Sockerless.
    This spec is the single source of truth for Go types and Backend interface generation.
  version: "1.44"
  license:
    name: Apache-2.0

# Custom extensions used by the code generator:
#   x-sockerless-go-method:       Backend interface method name
#   x-sockerless-go-args:         Path/query param → method arg mapping
#   x-sockerless-go-returns:      Return type override (e.g. io.ReadCloser)
#   x-sockerless-streaming:       Response is a stream (true/false)
#   x-sockerless-upgrade:         HTTP upgrade endpoint (attach, exec start)
#   x-sockerless-state-transition: Container state machine edge
#   x-sockerless-events:          Events emitted on success
#   x-sockerless-preconditions:   Required state before operation
#   x-sockerless-go-type:         Go type override for a schema field
#   x-sockerless-go-embeds:       Embedded struct types
#   x-sockerless-go-pointer:      Field should be a pointer type
#   x-sockerless-json-tag:        JSON tag override (when Go name differs from JSON)
#   x-sockerless-cloud-operations: Cloud API mapping per backend (documentation)
#   x-sockerless-extensions:      Fields that are Sockerless-specific (not in Docker spec)
paths:
...
  /events:
    get:
      operationId: SystemEvents
      x-sockerless-go-method: SystemEvents
      x-sockerless-streaming: true
      summary: Monitor events
      parameters:
        - name: since
          in: query
          schema: { type: string }
        - name: until
          in: query
          schema: { type: string }
        - name: filters
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Event stream
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
...
  /containers/create:
    post:
      operationId: ContainerCreate
      x-sockerless-go-method: ContainerCreate
      x-sockerless-state-transition:
        from: []
        to: created
      x-sockerless-preconditions: []
      x-sockerless-events:
        - { type: container, action: create }
      x-sockerless-cloud-operations:
        ecs: RegisterTaskDefinition
        lambda: CreateFunction
        cloudrun: services.create
        gcf: functions.create
        aca: containerApps.createOrUpdate
        azf: sites.createOrUpdate
      summary: Create a container
      parameters:
        - name: name
          in: query
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContainerCreateRequest"
      responses:
        "201":
          description: Container created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContainerCreateResponse"
  /containers/json:
    get:
      operationId: ContainerList
      x-sockerless-go-method: ContainerList
      summary: List containers
      parameters:
        - name: all
          in: query
          schema: { type: boolean }
        - name: limit
          in: query
          schema: { type: integer }
        - name: filters
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Container list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContainerSummary"
  /containers/{id}/json:
    get:
      operationId: ContainerInspect
      x-sockerless-go-method: ContainerInspect
      x-sockerless-go-args: [id]
      summary: Inspect a container
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Container details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Container"
  /containers/{id}/start:
    post:
      operationId: ContainerStart
      x-sockerless-go-method: ContainerStart
      x-sockerless-go-args: [id]
      x-sockerless-state-transition:
        from: [created, exited]
        to: running
      x-sockerless-preconditions: [container-exists, state-in(created,exited)]
      x-sockerless-events:
        - { type: container, action: start }
      x-sockerless-cloud-operations:
        ecs: RunTask
        lambda: Invoke (async)
        cloudrun: services.replaceService (traffic 100%)
        gcf: functions.patch (status ACTIVE)
        aca: containerApps.start
        azf: sites.start
      summary: Start a container
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Container started
  /containers/{id}/stop:
    post:
      operationId: ContainerStop
      x-sockerless-go-method: ContainerStop
      x-sockerless-go-args: [id]
      x-sockerless-state-transition:
        from: [running, paused]
        to: exited
      x-sockerless-preconditions: [container-exists, state-in(running,paused)]
      x-sockerless-events:
        - { type: container, action: die }
        - { type: container, action: stop }
      x-sockerless-cloud-operations:
        ecs: StopTask
        lambda: (no-op, execution completes)
        cloudrun: services.replaceService (traffic 0%)
        gcf: functions.patch (status OFFLINE)
        aca: containerApps.stop
        azf: sites.stop
      summary: Stop a container
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - name: t
          in: query
          schema: { type: integer }
      responses:
        "204":
          description: Container stopped
  /containers/{id}/restart:
    post:
      operationId: ContainerRestart
      x-sockerless-go-method: ContainerRestart
      x-sockerless-go-args: [id]
      x-sockerless-state-transition:
        from: [running, paused, exited, created]
        to: running
      x-sockerless-preconditions: [container-exists]
      x-sockerless-events:
        - { type: container, action: die }
        - { type: container, action: stop }
        - { type: container, action: start }
        - { type: container, action: restart }
      summary: Restart a container
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - name: t
          in: query
          schema: { type: integer }
      responses:
        "204":
          description: Container restarted
  /containers/{id}/kill:
    post:
      operationId: ContainerKill
      x-sockerless-go-method: ContainerKill
      x-sockerless-go-args: [id]
      x-sockerless-state-transition:
        from: [running, paused]
        to: exited
      x-sockerless-preconditions: [container-exists, state-in(running,paused)]
      x-sockerless-events:
        - { type: container, action: kill }
        - { type: container, action: die }
      summary: Kill a container
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - name: signal
          in: query
          schema: { type: string }
      responses:
        "204":
          description: Container killed
  /containers/{id}:
    delete:
      operationId: ContainerRemove
      x-sockerless-go-method: ContainerRemove
      x-sockerless-go-args: [id]
      x-sockerless-state-transition:
        from: [created, exited]
        to: removed
        force-from: [created, exited, running, paused]
      x-sockerless-preconditions: [container-exists, state-in(created,exited)]
      x-sockerless-events:
        - { type: container, action: destroy }
      x-sockerless-cloud-operations:
        ecs: DeregisterTaskDefinition + StopTask
        lambda: DeleteFunction
        cloudrun: services.delete
        gcf: functions.delete
        aca: containerApps.delete
        azf: sites.delete
      summary: Remove a container
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - name: force
          in: query
          schema: { type: boolean }
      responses:
        "204":
          description: Container removed
  /containers/{id}/logs:
    get:
      operationId: ContainerLogs
      x-sockerless-go-method: ContainerLogs
      x-sockerless-go-args: [id]
      x-sockerless-streaming: true
      x-sockerless-go-returns: io.ReadCloser
      summary: Get container logs
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - name: stdout
          in: query
          schema: { type: boolean }
        - name: stderr
          in: query
          schema: { type: boolean }
        - name: follow
          in: query
          schema: { type: boolean }
        - name: timestamps
          in: query
          schema: { type: boolean }
        - name: tail
          in: query
          schema: { type: string }
        - name: since
          in: query
          schema: { type: string }
        - name: until
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Log stream
  /containers/{id}/wait:
    post:
      operationId: ContainerWait
      x-sockerless-go-method: ContainerWait
      x-sockerless-go-args: [id]
      summary: Wait for a container
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - name: condition
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Wait response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContainerWaitResponse"
  /containers/{id}/attach:
    post:
      operationId: ContainerAttach
      x-sockerless-go-method: ContainerAttach
      x-sockerless-go-args: [id]
      x-sockerless-upgrade: true
      x-sockerless-go-returns: io.ReadWriteCloser
      summary: Attach to a container
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - name: stream
          in: query
          schema: { type: boolean }
        - name: stdin
          in: query
          schema: { type: boolean }
        - name: stdout
          in: query
          schema: { type: boolean }
        - name: stderr
          in: query
          schema: { type: boolean }
        - name: detachKeys
          in: query
          schema: { type: string }
        - name: logs
          in: query
          schema: { type: boolean }
      responses:
        "101":
          description: Upgraded to TCP
  /containers/{id}/resize:
    post:
      operationId: ContainerResize
      x-sockerless-go-method: ContainerResize
      x-sockerless-go-args: [id]
      summary: Resize a container TTY
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - name: h
          in: query
          schema: { type: integer }
        - name: w
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: Resized
  /containers/{id}/top:
    get:
      operationId: ContainerTop
      x-sockerless-go-method: ContainerTop
      x-sockerless-go-args: [id]
      summary: List processes in a container
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - name: ps_args
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Process list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContainerTopResponse"
  /containers/{id}/stats:
    get:
      operationId: ContainerStats
      x-sockerless-go-method: ContainerStats
      x-sockerless-go-args: [id]
      x-sockerless-streaming: true
      x-sockerless-go-returns: io.ReadCloser
      summary: Get container stats
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - name: stream
          in: query
          schema: { type: boolean }
      responses:
        "200":
          description: Stats stream
  /containers/{id}/rename:
    post:
      operationId: ContainerRename
      x-sockerless-go-method: ContainerRename
      x-sockerless-go-args: [id]
      x-sockerless-events:
        - { type: container, action: rename }
      summary: Rename a container
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - name: name
          in: query
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Renamed
  /containers/{id}/pause:
    post:
      operationId: ContainerPause
      x-sockerless-go-method: ContainerPause
      x-sockerless-go-args: [id]
      x-sockerless-state-transition:
        from: [running]
        to: paused
      x-sockerless-preconditions: [container-exists, state-in(running)]
      x-sockerless-events:
        - { type: container, action: pause }
      summary: Pause a container
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Paused
  /containers/{id}/unpause:
    post:
      operationId: ContainerUnpause
      x-sockerless-go-method: ContainerUnpause
      x-sockerless-go-args: [id]
      x-sockerless-state-transition:
        from: [paused]
        to: running
      x-sockerless-preconditions: [container-exists, state-in(paused)]
      x-sockerless-events:
        - { type: container, action: unpause }
      summary: Unpause a container
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Unpaused
  /containers/prune:
    post:
      operationId: ContainerPrune
      x-sockerless-go-method: ContainerPrune
      x-sockerless-events:
        - { type: container, action: destroy }
      summary: Delete stopped containers
      parameters:
        - name: filters
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Prune response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContainerPruneResponse"
  /containers/{id}/archive:
    put:
      operationId: ContainerPutArchive
      x-sockerless-go-method: ContainerPutArchive
      x-sockerless-go-args: [id]
      summary: Extract an archive of files into a container
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - name: path
          in: query
          required: true
          schema: { type: string }
        - name: noOverwriteDirNonDir
          in: query
          schema: { type: boolean }
      requestBody:
        content:
          application/x-tar:
            schema:
              type: string
              format: binary
      responses:
        "200":
          description: Archive extracted
    head:
      operationId: ContainerStatPath
      x-sockerless-go-method: ContainerStatPath
      x-sockerless-go-args: [id]
      summary: Get info about files in a container
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - name: path
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: File stat
    get:
      operationId: ContainerGetArchive
      x-sockerless-go-method: ContainerGetArchive
      x-sockerless-go-args: [id]
      summary: Get an archive of a filesystem resource in a container
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - name: path
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Archive stream
  /containers/{id}/changes:
    get:
      operationId: ContainerChanges
      x-sockerless-go-method: ContainerChanges
      x-sockerless-go-args: [id]
      summary: Get changes on a container filesystem
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Filesystem changes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContainerChangeItem"
  /containers/{id}/export:
    get:
      operationId: ContainerExport
      x-sockerless-go-method: ContainerExport
      x-sockerless-go-args: [id]
      x-sockerless-streaming: true
      x-sockerless-go-returns: io.ReadCloser
      summary: Export a container
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Tarball stream
  /containers/{id}/update:
    post:
      operationId: ContainerUpdate
      x-sockerless-go-method: ContainerUpdate
      x-sockerless-go-args: [id]
      x-sockerless-events:
        - { type: container, action: update }
      summary: Update a container
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContainerUpdateRequest"
      responses:
        "200":
          description: Update response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContainerUpdateResponse"

  # ── Exec ─────────────────────────────────────────────────────────────
  /containers/{id}/exec:
    post:
      operationId: ExecCreate
      x-sockerless-go-method: ExecCreate
      x-sockerless-go-args: [id]
      summary: Create an exec instance
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExecCreateRequest"
      responses:
        "201":
          description: Exec created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExecCreateResponse"
  /exec/{id}/start:
    post:
      operationId: ExecStart
      x-sockerless-go-method: ExecStart
      x-sockerless-go-args: [id]
      x-sockerless-upgrade: true
      x-sockerless-go-returns: io.ReadWriteCloser
      summary: Start an exec instance
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExecStartRequest"
      responses:
        "200":
          description: Exec output stream
  /exec/{id}/json:
    get:
      operationId: ExecInspect
      x-sockerless-go-method: ExecInspect
      x-sockerless-go-args: [id]
      summary: Inspect an exec instance
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Exec details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExecInstance"
  /exec/{id}/resize:
    post:
      operationId: ExecResize
      x-sockerless-go-method: ExecResize
      x-sockerless-go-args: [id]
      summary: Resize an exec instance TTY
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - name: h
          in: query
          schema: { type: integer }
        - name: w
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: Resized

  # ── Images ───────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ExecCreateResponse:
      type: object
      x-sockerless-docker-name: IdResponse
      properties:
        Id:
          type: string
...
    Event:
      type: object
      x-sockerless-docker-name: EventMessage
      properties:
        Type:
          type: string
        Action:
          type: string
        Actor:
          $ref: "#/components/schemas/EventActor"
        scope:
          type: string
        time:
          type: integer
          format: int64
        timeNano:
          type: integer
          format: int64
...